resume_agent/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
spacy==3.8.10
sentence-transformers==3.4.1
pdfplumber==0.11.5
python-docx==1.1.2
numpy==2.2.3
python-multipart==0.0.20
